requirements.txt
anthropic
python-dotenv
pandas
pytest
streamlit
pymongo
bcrypt
requests
Pillow
